Added OBJ export
From Discord:

I've just modified the texture swapper script to add a Dump to Obj
file button for the selected clothing item. (We should set up a github
:wink: )
This will dump all the meshes on the skin to .obj files in the root -
it uses the built in ObjExporter.
As we don't have write access.